fix(video-player): prevent minuscule thumbnails (#9583)
### Related Ticket(s)

Resolves #9365 

### Description

The default width for `DDSVideoPlayer` thumbnails seems to have been accidentally set to `3px`, resulting in tiny images that appeared as a blur when stretched to fill the container. I've reset this default width to `655px`, the size they were before the accidental change.

While debugging, I noticed the `tabs-extended-media` Storybook story was not rendering videos. I fixed this by adding a missing import to the `content-item-horizontal/index.ts` file.

### Changelog

**Changed**

- Restores the default width for video player thumbnails to a sensible size (`655px`).
- Fixes video rendering when using `content-item-horizontal` component.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->