fix(masthead): avoid inline styles for CSP compliance (#9605)
## Related Ticket(s)

Closes #9599 

## Description

Due to the content-security-policy header, browsers will begin to [block inline scripts and styles](https://content-security-policy.com/examples/allow-inline-style/). To keep the masthead's `<top-nav>` component usable when the number of items would over flow the container, we need a way to handle offset styles _dynamically_ without using inline styles or regular `<style>` elements on the page.

This PR adds a new mixin that creates a `<style>` element with the necessary `nonce` attribute, if available. The mixin provides the below functionality

### On ConnectedCallback

When the component connects, it checks `window.nonce` for the nonce string for this exact pageload, and stores it as `this._nonce` for later use

### Render functions

Components implmenting this mixin should _not_ have their own render function, but should instead use the abstract `renderContents()` method. This mixin renders the dynamic stylesheet along with the returned template from renderContents(). If a component _must_ implement `render()`, it must also run `renderDynamicStyles()` within that render function.

### Style getter/setter methods

- `setStyleBySelector(selectorString, styleProperty, styleValue) { ... }`
  "upserts" a style rule, either creating it for the first time or overriding the style for the given selector
  
- `getStyleBySelector(selectorString, styleProperty) { ... }`

### The Stylesheet

The rendered `<style>` element will not update it's contents, but because we're using CSSOM methods, the styles _do_ apply, and they _do_ cause change on the page. Users should never attempt to set `innerHTML` of the element, as that will cause the element to "reset" and discard all programatic changes made using the setter method above.



### Changelog

**Changed**

- Moved DDSTopNav inline styles to an adoptedStylesheet

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->