fix(left-nav): fix RTL menu attachment and movement (#9820)
### Related Ticket(s)

[ADCMS-2656](https://jsw.ibm.com/browse/ADCMS-2656)

### Description

Fixes the `dds-left-nav` component in RTL pages. Previously the left-nav stuck to the left side of the screen, left the overlay cutout on the left side, and animated the menu sections in a left-to-right fashion. This PR updates the styles to use CSS logical properties - specifically `inset-inline-start` to use the page's writing mode to dictate how the mobile nav should behave.

### Changelog

**Changed**

- fixed `dds-left-nav` in RTL pages