fix(back-to-top): trigger manage visibility on body height changes (#9900)
### Related Ticket(s)

Closes: https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/8834

[ADCMS-2857](https://jsw.ibm.com/browse/ADCMS-2857)

### Description

Reproducing this bug was a little tricky within Storybook. What I did was add a story for the Back to Top component called "Default Faux Paged". Initially with the same contents as the "Default" story. At the bottom of the page, there is a button, which when clicked will replace the page contents with other contents, speifically _less_ content, where no Back to Top button would be required at a typical desktop break.

The core of the issue was that, although we have a ResizeObserver to watch changes in the body element height, we don't make use of it to show / hide the Back to Top button when the height changes. I've adjusted the logic and renamed a couple of things, to hopefully make the code more readable.

### Testing instructions

- [ ] Navigate to the Default Faux Paged story under the Back to Top component at a Desktop breakpoint
- [ ] Scroll down the page, until you see the "Next page" button. As you scroll the Back to Top button will show up per usual.
- [ ] Click Next Page. The contents of the page should be replaced.
- [ ] The Back to Top button should dissapear, and not persist. This will depend on your viewport height.
- [ ] Scroll the new content up and down, the Back to Top button should not appear.

### Changelog

**New**

- Back to Top / Default Faux Paged story

**Changed**

- Updated back to top logic to cover body content change scenarios

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->