[Cloud-Masthead]: Implement lazy load and DOM pruning from masthead v2 on cloud masthead (#10085)
### Related Ticket(s)

Resolves #10056

### Description

Adds the performance enhancements that were implemented for Masthead v2 to the Cloud Masthead, specifically:

* https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/9595
* https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/9602

The first change (quoting @andy-blum [here](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/9595)): Initializes the masthead-container with either the `left-nav` or the `top-nav` depending on the width of the viewport. Additionally swaps one for the other when the viewport crosses the 960px boundary. This should effectively cut the number of masthead nodes in half with very little other changes.

The second change (again quoting @andy-blum [here](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/9602)): Currently the Masthead loads both the SideNav and Megamenu components at the same time, even if they haven't been opened yet. This causes the amount of nodes to increase considerably right away. In order to mitigate this, a fast solution is to dynamically load the required components *only* when the user has interacted with the menus.

This means that the mobile version wouldn't be loading the Megamenu, and the desktop wouldn't load SideNav until/if required. 

Another optimization that was done ensures that the Megamenu component wouldn't get loaded into the DOM unless a user has clicked on the menu beforehand. Then and only then, the megamenu would be loaded in. Once the menu is closed, the component gets removed from the DOM to ensure the total node count doesn't shoot up.

Note that, due to the structure of the Masthead and Cloud Masthead, these changes are not entirely isolated to the Cloud Masthead alone. Specifically, the Masthead is responsible for mega menu rendering, thus the Masthead implements lazy loading of the megamenu which then is inherited by the Cloud Masthead.

### Testing

Testing should cover regression tests on both the Masthead and Cloud Masthead components.

### Changelog

**New**

- Adds support for dynamic imports
- Utliizes dynamic imports to lazy load left nav and megamenu scripts that are only necessary when the user expands the menu.
- Only renders left nav custom elements when we're at the mobile break
- Only renders megamenu custom elements when user expands the menu