fix(image-with-caption): adjust image button background color (#10279)
### Related Ticket(s)

https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/10250
https://jsw.ibm.com/browse/ADCMS-3105

### Description

Sets the background style to transparent for the `<button>` element used in the shadow DOM of the `<dds-image>` and `<dds-image-with-caption>` components, such that when used with a transparent image (SVG, transparent gif/png), the browser's default button grey does not show through.

Note that the original issue was reported for `<dds-image-with-caption>`, but the issue equally effects `<dds-image>` component as well. Note also that the `<dds-image-with-caption>` button was recently deprecated, it's features merged with `<dds-image>`. Use of `<dds-image-with-caption>` should be replaced by `<dds-image>` (see https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/8839). The deprecation is documented in the [Storybook README for the Image component](https://carbon-design-system.github.io/carbon-for-ibm-dotcom/canary/web-components/?path=/docs/components-image--default#image-with-caption).

### Testing Instructions

- [ ] Open Storybook to the Image > Default component. 
- [ ] For the "Default image" Storybook knob, choose "SVG". A transparent SVG is used as the image for the component. 
- [ ] Select Lightbox as well from the Storybook knobs
- [ ] Note the browser default button color does not show through.

### Changelog

**Changed**

- Added `background-color: transparent` to the button styles for the `<dds-image>` component when in lightbox mode.