fix(leadspace): fix heading width for the leadspace (#10303)
### Related Ticket(s)

https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/10262
https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/10261

https://jsw.ibm.com/browse/ADCMS-3154

### Description

Fixes a couple of Leadspace issues. First issue it to do with the heading width, it's off according to the design specs. This brings it in line with the spec for both the medium and large breakpoints. Second issue is when the background image shifts from showing on its own line to inline with the heading and description. It now only shifts inline at the large breakpoint.

### Changelog

**Changed**

- Leadspace background now only shifts to inline with headeing and description at the large breakpoint
- Spacing adjustments for the Leadspace title width at medium and large breakpoints per [design specs](https://ibm.ent.box.com/folder/94761007991?s=d9jey3rlfxqmxu059f0x9dtcx7f5ehuh).