Adjust the handling of videoThumbnailUrl property to avoid small images from Kaltura API (#10541)
### Related Ticket(s)

https://jsw.ibm.com/browse/ADCMS-3370
https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/10522

### Description

Adjust the handling of videoThumbnailUrl property on CTA's that use the `VideoCTAMixin`. This should avoid the bug that we are currently seeing in production wher the CTA component `offsetWidth` is not yet computed, resulting in a small thumbnail width that is blurry when stretched to fit.

There isn't a great way to test the specific condition being observed [in production right now](https://ibm.com/products/envizi/resources). This appears to be a race condition, such that when the Kaltura API is leverged to fetch a thumbnail image, the width determined for the videos container via `offsetWidth` is not available and returns 0. The changes here should more reliably set the thumbnail after the `offsetWidth` is available. This borrows the logic from a [similar fix here](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/pull/7914/files#diff-c16cc0206fb8b5202bfa8ddd7900037a71fb8b30235347e4c33ece727f8d8860R233-R235).

### Changelog

**Changed**

- `dds-video-cta-composite` no longer manages the`videoThumbnailUrl`, that handling is now encapsulated in the VideoCTAMixin, on the component where the `videoThumbnailUrl` is actually used.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->

### Testing

- [ ] Navigate to the Card Group > Default story. Select Add Media from the Knobs. The Video CTA cards should render an image with a width that matches the width of the card.
- [ ] Regression: Navigate to the CTA > Default story. Select CTA Type of "Card" and CTA style of "Video (video)". The Video CTA card should render with an image with a width that matches the width of the card.
- [ ] Regression: Navigate to the Card in Card > Default story. Select "video". The Video CTA card should render with an image with a width tha matches the width of the card.