fix(lightbox-video-player): use more reliable reference to video player (#10878)
### Related Ticket(s)

none

### Description

`<dds-video-cta-container>` can handle multiple Kaltura video embed requests within the same modal/lightbox by slotting a `<div>` per video into the same `<dds-lightbox-video-player>`. When the modal is activated by user interaction, the current video ID is sent along, and the `<dds-lightbox-video-player-composite>` sets `hidden` on the `<div>`s that don't have a matching ID. This works great when there is only one `<dds-video-cta-container>`.

The problems start when multiple video containers are used throughout the document. Each creates some number modal render root elements depending on the structure of the content, and the lightbox composite uses `document.querySelector('dds-lightbox-video-player')` to find the video player when trying to hide embedded Kaltura videos that are not active. As we know, `querySelector` returns the first matching element, and since we've established there may be multiple of these video players in various modal render roots, sometimes it found the wrong one and failed to hide non-active videos.

Luckily, the fix was pretty simple. Lightbox composites already have an internal reference to their video player, so I swapped out the document query to use that reference, which _should_ guarantee the composite is looking in the right place when hiding inactive videos.

### Changelog

**Changed**

- Ensure `<dds-lightbox-video-player-composite>` hides inactive videos by using an internal reference to modal's video player instead of querying for it from the document root.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->