[Filter panel]: Reduce Filter Panel DOM node counts (#11057)
### Related Ticket(s)

Closes #10669 
Closes https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/9040
[Jira ADCMS-4089](https://jsw.ibm.com/browse/ADCMS-4089)
[Jira ADCMS-4084](https://jsw.ibm.com/browse/ADCMS-4084)

### Description

Reduces DOM node counts for the filter panel by only rendering the mobile or destop experience according to viewport. In order to acheive this, does a bit of refactoring how the heading is handled. Before the desktop and mobile layouts were coupled by only the mobile layout rendering the heading slot.

### Changelog

**Changed**

- Filter panel only renders desktop / mobile layout per the viewport

### Testing instructions

- [ ] Regression testing the Filter Panel
- [ ] Inspect the DOM. Only markup for desktop or mobile should be present, never both.