fix(pricing-table): cell items were being horizontally squished into … (#11002)
…the cell

### Related Ticket(s)

https://jsw.ibm.com/browse/ADCMS-3714

### Description

In a Pricing Table, for the Table Row when the user choses to "Add Cell Content" for cells to contain multiple paragraphs, or combination of paragraph text, Tag Group tags, or Links, when the page is published the Pricing table is displaying this added cell content distributed horizontally squished into the cell. The Added Cell Content is displaying to the right of the initial cell content in the cell.
![Screenshot 2023-10-04 at 11 29 52](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/assets/7583437/6f189fac-868c-4e55-9b12-b6cb15d71e8b)

I've changed the flex direction to column inside of the cell contents so that all items will sit one below the other.